+++ vscode/src/extension.ts
@@ -21,6 +21,8 @@ import listener from "./core/listeners";
 import { ChangelogPanel } from "./panels/ChangelogPanel";
 import { WelcomePagePanel } from "./panels/WelcomePanel";
 import { ExtensionStorage } from "./storage";
+import { disableLockFileParsing } from "./commands/disableLockFileParsing";
+import { enableLockFileParsing } from "./commands/enableLockFileParsing";
 
 export var Logger: OutputChannel;
 
@@ -83,6 +85,8 @@ export function activate(context: ExtensionContext) {
   context.subscriptions.push(updateAll);
   context.subscriptions.push(generateVulnerabilityReport(context));
   context.subscriptions.push(generateCurrentVulnReport(context));
+  context.subscriptions.push(disableLockFileParsing);
+  context.subscriptions.push(enableLockFileParsing);
 }
 
 export function deactivate() {